+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fix the issue when fee/fine is partially paid, then refunded, User Details show the full amount of the fee/fine as refunded. Refs UIU-2455.
 * Fix the issue when a fee/fine is refunded due to a CLAIMED RETURNED, the refund amount does not appear in User Details. Refs UIU-2469.
 * Correctly check permissions for accounts routes. Refs UIU-2474.
+* Refactor from `<SafeHTMLMessage>` to `<FormattedMessage>`. Refs UIU-2179.
 
 ## [7.0.1](https://github.com/folio-org/ui-users/tree/v7.0.1) (2021-10-07)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v6.1.0...v7.0.1)